app/requirements.txt
azure-common==1.1.6
azure-batch==3.0.0
azure-storage==0.34.2
django==1.11.4
psycopg2==2.7.3
requests==2.18.3